refactor(APISIMP-PUBS-LIST-REST-DELETE): delete PublicationsListRest 410 tombstone, extract absoluteUrl helper
The per-kind deprecated alias GET /v2/{kind}/{appId}/publications has been
a 410 Gone tombstone since fire-505. Deletes the class and its test (5 tests
verifying the 410 behaviour). The sole live logic was the package-private
absoluteUrl() helper consumed by FlatPublicationsRest; extracted to a new
PublishRestUtils utility class to cut the dependency before deleting the
tombstone.

- Delete PublicationsListRest.java (410 tombstone, no remaining callers)
- Delete PublicationsListRestTest.java (5 410-tombstone tests)
- Add PublishRestUtils.java with absoluteUrl() static helper
- FlatPublicationsRest: call PublishRestUtils.absoluteUrl() instead
- FlatPublicationsRestTest: remove stale PublicationsListRestTest javadoc ref
- usePublicationStatus.ts: update stale comment referencing old endpoint
- aidocs/16: add APISIMP-PUBS-LIST-REST-DELETE + 4 new sweep-fire-563 rows;
  update PR #2499 SHA and PR #2500 reference
